+++ src/datamigration/azext_datamigration/vendored_sdks/datamigration/aio/operations/_database_migrations_sql_db_operations.py
@@ -20,8 +20,8 @@ from ... import models
 T = TypeVar('T')
 ClsType = Optional[Callable[[PipelineResponse[HttpRequest, AsyncHttpResponse], T, Dict[str, Any]], Any]]
 
-class DatabaseMigrationsSqlMiOperations:
-    """DatabaseMigrationsSqlMiOperations async operations.
+class DatabaseMigrationsSqlDbOperations:
+    """DatabaseMigrationsSqlDbOperations async operations.
 
     You should not instantiate this class directly. Instead, you should create a Client instance that
     instantiates it for you and attaches it as an attribute.
@@ -45,19 +45,19 @@ class DatabaseMigrationsSqlMiOperations:
     async def get(
         self,
         resource_group_name: str,
-        managed_instance_name: str,
+        sql_db_instance_name: str,
         target_db_name: str,
         migration_operation_id: Optional[str] = None,
         expand: Optional[str] = None,
         **kwargs
-    ) -> "models.DatabaseMigrationSqlMi":
-        """Retrieve the specified database migration for a given SQL Managed Instance.
+    ) -> "models.DatabaseMigrationSqlDb":
+        """Retrieve the Database Migration resource.
 
         :param resource_group_name: Name of the resource group that contains the resource. You can
          obtain this value from the Azure Resource Manager API or the portal.
         :type resource_group_name: str
-        :param managed_instance_name:
-        :type managed_instance_name: str
+        :param sql_db_instance_name:
+        :type sql_db_instance_name: str
         :param target_db_name: The name of the target database.
         :type target_db_name: str
         :param migration_operation_id: Optional migration operation ID. If this is provided, then
@@ -67,11 +67,11 @@ class DatabaseMigrationsSqlMiOperations:
         :param expand: Complete migration details be included in the response.
         :type expand: str
         :keyword callable cls: A custom type or function that will be passed the direct response
-        :return: DatabaseMigrationSqlMi, or the result of cls(response)
-        :rtype: ~azure.mgmt.datamigration.models.DatabaseMigrationSqlMi
+        :return: DatabaseMigrationSqlDb, or the result of cls(response)
+        :rtype: ~azure.mgmt.datamigration.models.DatabaseMigrationSqlDb
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        cls = kwargs.pop('cls', None)  # type: ClsType["models.DatabaseMigrationSqlMi"]
+        cls = kwargs.pop('cls', None)  # type: ClsType["models.DatabaseMigrationSqlDb"]
         error_map = {
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
@@ -83,7 +83,7 @@ class DatabaseMigrationsSqlMiOperations:
         url = self.get.metadata['url']  # type: ignore
         path_format_arguments = {
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'managedInstanceName': self._serialize.url("managed_instance_name", managed_instance_name, 'str'),
+            'sqlDbInstanceName': self._serialize.url("sql_db_instance_name", sql_db_instance_name, 'str'),
             'targetDbName': self._serialize.url("target_db_name", target_db_name, 'str'),
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
         }
@@ -109,23 +109,23 @@ class DatabaseMigrationsSqlMiOperations:
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response, error_format=ARMErrorFormat)
 
-        deserialized = self._deserialize('DatabaseMigrationSqlMi', pipeline_response)
+        deserialized = self._deserialize('DatabaseMigrationSqlDb', pipeline_response)
 
         if cls:
             return cls(pipeline_response, deserialized, {})
 
         return deserialized
-    get.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/managedInstances/{managedInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}'}  # type: ignore
+    get.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/servers/{sqlDbInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}'}  # type: ignore
 
     async def _create_or_update_initial(
         self,
         resource_group_name: str,
-        managed_instance_name: str,
+        sql_db_instance_name: str,
         target_db_name: str,
-        parameters: "models.DatabaseMigrationSqlMi",
+        parameters: "models.DatabaseMigrationSqlDb",
         **kwargs
-    ) -> "models.DatabaseMigrationSqlMi":
-        cls = kwargs.pop('cls', None)  # type: ClsType["models.DatabaseMigrationSqlMi"]
+    ) -> "models.DatabaseMigrationSqlDb":
+        cls = kwargs.pop('cls', None)  # type: ClsType["models.DatabaseMigrationSqlDb"]
         error_map = {
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
@@ -138,7 +138,7 @@ class DatabaseMigrationsSqlMiOperations:
         url = self._create_or_update_initial.metadata['url']  # type: ignore
         path_format_arguments = {
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'managedInstanceName': self._serialize.url("managed_instance_name", managed_instance_name, 'str'),
+            'sqlDbInstanceName': self._serialize.url("sql_db_instance_name", sql_db_instance_name, 'str'),
             'targetDbName': self._serialize.url("target_db_name", target_db_name, 'str'),
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
         }
@@ -154,7 +154,7 @@ class DatabaseMigrationsSqlMiOperations:
         header_parameters['Accept'] = self._serialize.header("accept", accept, 'str')
 
         body_content_kwargs = {}  # type: Dict[str, Any]
-        body_content = self._serialize.body(parameters, 'DatabaseMigrationSqlMi')
+        body_content = self._serialize.body(parameters, 'DatabaseMigrationSqlDb')
         body_content_kwargs['content'] = body_content
         request = self._client.put(url, query_parameters, header_parameters, **body_content_kwargs)
         pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
@@ -165,48 +165,48 @@ class DatabaseMigrationsSqlMiOperations:
             raise HttpResponseError(response=response, error_format=ARMErrorFormat)
 
         if response.status_code == 200:
-            deserialized = self._deserialize('DatabaseMigrationSqlMi', pipeline_response)
+            deserialized = self._deserialize('DatabaseMigrationSqlDb', pipeline_response)
 
         if response.status_code == 201:
-            deserialized = self._deserialize('DatabaseMigrationSqlMi', pipeline_response)
+            deserialized = self._deserialize('DatabaseMigrationSqlDb', pipeline_response)
 
         if cls:
             return cls(pipeline_response, deserialized, {})
 
         return deserialized
-    _create_or_update_initial.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/managedInstances/{managedInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}'}  # type: ignore
+    _create_or_update_initial.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/servers/{sqlDbInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}'}  # type: ignore
 
     async def begin_create_or_update(
         self,
         resource_group_name: str,
-        managed_instance_name: str,
+        sql_db_instance_name: str,
         target_db_name: str,
-        parameters: "models.DatabaseMigrationSqlMi",
+        parameters: "models.DatabaseMigrationSqlDb",
         **kwargs
-    ) -> AsyncLROPoller["models.DatabaseMigrationSqlMi"]:
-        """Create a new database migration to a given SQL Managed Instance.
+    ) -> AsyncLROPoller["models.DatabaseMigrationSqlDb"]:
+        """Create a new database migration to a given SQL DB.
 
         :param resource_group_name: Name of the resource group that contains the resource. You can
          obtain this value from the Azure Resource Manager API or the portal.
         :type resource_group_name: str
-        :param managed_instance_name:
-        :type managed_instance_name: str
+        :param sql_db_instance_name:
+        :type sql_db_instance_name: str
         :param target_db_name: The name of the target database.
         :type target_db_name: str
-        :param parameters: Details of SqlMigrationService resource.
-        :type parameters: ~azure.mgmt.datamigration.models.DatabaseMigrationSqlMi
+        :param parameters: Details of Sql Db migration resource.
+        :type parameters: ~azure.mgmt.datamigration.models.DatabaseMigrationSqlDb
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
         :keyword polling: True for ARMPolling, False for no polling, or a
          polling object for personal polling strategy
         :paramtype polling: bool or ~azure.core.polling.AsyncPollingMethod
         :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
-        :return: An instance of AsyncLROPoller that returns either DatabaseMigrationSqlMi or the result of cls(response)
-        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.mgmt.datamigration.models.DatabaseMigrationSqlMi]
+        :return: An instance of AsyncLROPoller that returns either DatabaseMigrationSqlDb or the result of cls(response)
+        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.mgmt.datamigration.models.DatabaseMigrationSqlDb]
         :raises ~azure.core.exceptions.HttpResponseError:
         """
         polling = kwargs.pop('polling', True)  # type: Union[bool, AsyncPollingMethod]
-        cls = kwargs.pop('cls', None)  # type: ClsType["models.DatabaseMigrationSqlMi"]
+        cls = kwargs.pop('cls', None)  # type: ClsType["models.DatabaseMigrationSqlDb"]
         lro_delay = kwargs.pop(
             'polling_interval',
             self._config.polling_interval
@@ -215,7 +215,7 @@ class DatabaseMigrationsSqlMiOperations:
         if cont_token is None:
             raw_result = await self._create_or_update_initial(
                 resource_group_name=resource_group_name,
-                managed_instance_name=managed_instance_name,
+                sql_db_instance_name=sql_db_instance_name,
                 target_db_name=target_db_name,
                 parameters=parameters,
                 cls=lambda x,y,z: x,
@@ -226,7 +226,7 @@ class DatabaseMigrationsSqlMiOperations:
         kwargs.pop('content_type', None)
 
         def get_long_running_output(pipeline_response):
-            deserialized = self._deserialize('DatabaseMigrationSqlMi', pipeline_response)
+            deserialized = self._deserialize('DatabaseMigrationSqlDb', pipeline_response)
 
             if cls:
                 return cls(pipeline_response, deserialized, {})
@@ -234,7 +234,7 @@ class DatabaseMigrationsSqlMiOperations:
 
         path_format_arguments = {
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'managedInstanceName': self._serialize.url("managed_instance_name", managed_instance_name, 'str'),
+            'sqlDbInstanceName': self._serialize.url("sql_db_instance_name", sql_db_instance_name, 'str'),
             'targetDbName': self._serialize.url("target_db_name", target_db_name, 'str'),
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
         }
@@ -251,12 +251,134 @@ class DatabaseMigrationsSqlMiOperations:
             )
         else:
             return AsyncLROPoller(self._client, raw_result, get_long_running_output, polling_method)
-    begin_create_or_update.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/managedInstances/{managedInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}'}  # type: ignore
+    begin_create_or_update.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/servers/{sqlDbInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}'}  # type: ignore
+
+    async def _delete_initial(
+        self,
+        resource_group_name: str,
+        sql_db_instance_name: str,
+        target_db_name: str,
+        force: Optional[bool] = None,
+        **kwargs
+    ) -> None:
+        cls = kwargs.pop('cls', None)  # type: ClsType[None]
+        error_map = {
+            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
+        }
+        error_map.update(kwargs.pop('error_map', {}))
+        api_version = "2022-03-30-preview"
+
+        # Construct URL
+        url = self._delete_initial.metadata['url']  # type: ignore
+        path_format_arguments = {
+            'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
+            'sqlDbInstanceName': self._serialize.url("sql_db_instance_name", sql_db_instance_name, 'str'),
+            'targetDbName': self._serialize.url("target_db_name", target_db_name, 'str'),
+            'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
+        }
+        url = self._client.format_url(url, **path_format_arguments)
+
+        # Construct parameters
+        query_parameters = {}  # type: Dict[str, Any]
+        if force is not None:
+            query_parameters['force'] = self._serialize.query("force", force, 'bool')
+        query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
+
+        # Construct headers
+        header_parameters = {}  # type: Dict[str, Any]
+
+        request = self._client.delete(url, query_parameters, header_parameters)
+        pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
+        response = pipeline_response.http_response
+
+        if response.status_code not in [200, 202, 204]:
+            map_error(status_code=response.status_code, response=response, error_map=error_map)
+            raise HttpResponseError(response=response, error_format=ARMErrorFormat)
+
+        if cls:
+            return cls(pipeline_response, None, {})
+
+    _delete_initial.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/servers/{sqlDbInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}'}  # type: ignore
+
+    async def begin_delete(
+        self,
+        resource_group_name: str,
+        sql_db_instance_name: str,
+        target_db_name: str,
+        force: Optional[bool] = None,
+        **kwargs
+    ) -> AsyncLROPoller[None]:
+        """Delete Database Migration resource.
+
+        :param resource_group_name: Name of the resource group that contains the resource. You can
+         obtain this value from the Azure Resource Manager API or the portal.
+        :type resource_group_name: str
+        :param sql_db_instance_name:
+        :type sql_db_instance_name: str
+        :param target_db_name: The name of the target database.
+        :type target_db_name: str
+        :param force: Optional force delete boolean. If this is provided as true, migration will be
+         deleted even if active.
+        :type force: bool
+        :keyword callable cls: A custom type or function that will be passed the direct response
+        :keyword str continuation_token: A continuation token to restart a poller from a saved state.
+        :keyword polling: True for ARMPolling, False for no polling, or a
+         polling object for personal polling strategy
+        :paramtype polling: bool or ~azure.core.polling.AsyncPollingMethod
+        :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
+        :return: An instance of AsyncLROPoller that returns either None or the result of cls(response)
+        :rtype: ~azure.core.polling.AsyncLROPoller[None]
+        :raises ~azure.core.exceptions.HttpResponseError:
+        """
+        polling = kwargs.pop('polling', True)  # type: Union[bool, AsyncPollingMethod]
+        cls = kwargs.pop('cls', None)  # type: ClsType[None]
+        lro_delay = kwargs.pop(
+            'polling_interval',
+            self._config.polling_interval
+        )
+        cont_token = kwargs.pop('continuation_token', None)  # type: Optional[str]
+        if cont_token is None:
+            raw_result = await self._delete_initial(
+                resource_group_name=resource_group_name,
+                sql_db_instance_name=sql_db_instance_name,
+                target_db_name=target_db_name,
+                force=force,
+                cls=lambda x,y,z: x,
+                **kwargs
+            )
+
+        kwargs.pop('error_map', None)
+        kwargs.pop('content_type', None)
+
+        def get_long_running_output(pipeline_response):
+            if cls:
+                return cls(pipeline_response, None, {})
+
+        path_format_arguments = {
+            'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
+            'sqlDbInstanceName': self._serialize.url("sql_db_instance_name", sql_db_instance_name, 'str'),
+            'targetDbName': self._serialize.url("target_db_name", target_db_name, 'str'),
+            'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
+        }
+
+        if polling is True: polling_method = AsyncARMPolling(lro_delay, path_format_arguments=path_format_arguments,  **kwargs)
+        elif polling is False: polling_method = AsyncNoPolling()
+        else: polling_method = polling
+        if cont_token:
+            return AsyncLROPoller.from_continuation_token(
+                polling_method=polling_method,
+                continuation_token=cont_token,
+                client=self._client,
+                deserialization_callback=get_long_running_output
+            )
+        else:
+            return AsyncLROPoller(self._client, raw_result, get_long_running_output, polling_method)
+    begin_delete.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/servers/{sqlDbInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}'}  # type: ignore
 
     async def _cancel_initial(
         self,
         resource_group_name: str,
-        managed_instance_name: str,
+        sql_db_instance_name: str,
         target_db_name: str,
         parameters: "models.MigrationOperationInput",
         **kwargs
@@ -273,7 +395,7 @@ class DatabaseMigrationsSqlMiOperations:
         url = self._cancel_initial.metadata['url']  # type: ignore
         path_format_arguments = {
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'managedInstanceName': self._serialize.url("managed_instance_name", managed_instance_name, 'str'),
+            'sqlDbInstanceName': self._serialize.url("sql_db_instance_name", sql_db_instance_name, 'str'),
             'targetDbName': self._serialize.url("target_db_name", target_db_name, 'str'),
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
         }
@@ -301,23 +423,23 @@ class DatabaseMigrationsSqlMiOperations:
         if cls:
             return cls(pipeline_response, None, {})
 
-    _cancel_initial.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/managedInstances/{managedInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}/cancel'}  # type: ignore
+    _cancel_initial.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/servers/{sqlDbInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}/cancel'}  # type: ignore
 
     async def begin_cancel(
         self,
         resource_group_name: str,
-        managed_instance_name: str,
+        sql_db_instance_name: str,
         target_db_name: str,
         parameters: "models.MigrationOperationInput",
         **kwargs
     ) -> AsyncLROPoller[None]:
-        """Stop in-progress database migration to SQL Managed Instance.
+        """Stop on going migration for the database.
 
         :param resource_group_name: Name of the resource group that contains the resource. You can
          obtain this value from the Azure Resource Manager API or the portal.
         :type resource_group_name: str
-        :param managed_instance_name:
-        :type managed_instance_name: str
+        :param sql_db_instance_name:
+        :type sql_db_instance_name: str
         :param target_db_name: The name of the target database.
         :type target_db_name: str
         :param parameters: Required migration operation ID for which cancel will be initiated.
@@ -342,7 +464,7 @@ class DatabaseMigrationsSqlMiOperations:
         if cont_token is None:
             raw_result = await self._cancel_initial(
                 resource_group_name=resource_group_name,
-                managed_instance_name=managed_instance_name,
+                sql_db_instance_name=sql_db_instance_name,
                 target_db_name=target_db_name,
                 parameters=parameters,
                 cls=lambda x,y,z: x,
@@ -358,7 +480,7 @@ class DatabaseMigrationsSqlMiOperations:
 
         path_format_arguments = {
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'managedInstanceName': self._serialize.url("managed_instance_name", managed_instance_name, 'str'),
+            'sqlDbInstanceName': self._serialize.url("sql_db_instance_name", sql_db_instance_name, 'str'),
             'targetDbName': self._serialize.url("target_db_name", target_db_name, 'str'),
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
         }
@@ -375,128 +497,4 @@ class DatabaseMigrationsSqlMiOperations:
             )
         else:
             return AsyncLROPoller(self._client, raw_result, get_long_running_output, polling_method)
-    begin_cancel.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/managedInstances/{managedInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}/cancel'}  # type: ignore
-
-    async def _cutover_initial(
-        self,
-        resource_group_name: str,
-        managed_instance_name: str,
-        target_db_name: str,
-        parameters: "models.MigrationOperationInput",
-        **kwargs
-    ) -> None:
-        cls = kwargs.pop('cls', None)  # type: ClsType[None]
-        error_map = {
-            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
-        }
-        error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-03-30-preview"
-        content_type = kwargs.pop("content_type", "application/json")
-
-        # Construct URL
-        url = self._cutover_initial.metadata['url']  # type: ignore
-        path_format_arguments = {
-            'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'managedInstanceName': self._serialize.url("managed_instance_name", managed_instance_name, 'str'),
-            'targetDbName': self._serialize.url("target_db_name", target_db_name, 'str'),
-            'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
-        }
-        url = self._client.format_url(url, **path_format_arguments)
-
-        # Construct parameters
-        query_parameters = {}  # type: Dict[str, Any]
-        query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
-
-        # Construct headers
-        header_parameters = {}  # type: Dict[str, Any]
-        header_parameters['Content-Type'] = self._serialize.header("content_type", content_type, 'str')
-
-        body_content_kwargs = {}  # type: Dict[str, Any]
-        body_content = self._serialize.body(parameters, 'MigrationOperationInput')
-        body_content_kwargs['content'] = body_content
-        request = self._client.post(url, query_parameters, header_parameters, **body_content_kwargs)
-        pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
-        response = pipeline_response.http_response
-
-        if response.status_code not in [200, 202]:
-            map_error(status_code=response.status_code, response=response, error_map=error_map)
-            raise HttpResponseError(response=response, error_format=ARMErrorFormat)
-
-        if cls:
-            return cls(pipeline_response, None, {})
-
-    _cutover_initial.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/managedInstances/{managedInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}/cutover'}  # type: ignore
-
-    async def begin_cutover(
-        self,
-        resource_group_name: str,
-        managed_instance_name: str,
-        target_db_name: str,
-        parameters: "models.MigrationOperationInput",
-        **kwargs
-    ) -> AsyncLROPoller[None]:
-        """Initiate cutover for in-progress online database migration to SQL Managed Instance.
-
-        :param resource_group_name: Name of the resource group that contains the resource. You can
-         obtain this value from the Azure Resource Manager API or the portal.
-        :type resource_group_name: str
-        :param managed_instance_name:
-        :type managed_instance_name: str
-        :param target_db_name: The name of the target database.
-        :type target_db_name: str
-        :param parameters: Required migration operation ID for which cutover will be initiated.
-        :type parameters: ~azure.mgmt.datamigration.models.MigrationOperationInput
-        :keyword callable cls: A custom type or function that will be passed the direct response
-        :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: True for ARMPolling, False for no polling, or a
-         polling object for personal polling strategy
-        :paramtype polling: bool or ~azure.core.polling.AsyncPollingMethod
-        :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
-        :return: An instance of AsyncLROPoller that returns either None or the result of cls(response)
-        :rtype: ~azure.core.polling.AsyncLROPoller[None]
-        :raises ~azure.core.exceptions.HttpResponseError:
-        """
-        polling = kwargs.pop('polling', True)  # type: Union[bool, AsyncPollingMethod]
-        cls = kwargs.pop('cls', None)  # type: ClsType[None]
-        lro_delay = kwargs.pop(
-            'polling_interval',
-            self._config.polling_interval
-        )
-        cont_token = kwargs.pop('continuation_token', None)  # type: Optional[str]
-        if cont_token is None:
-            raw_result = await self._cutover_initial(
-                resource_group_name=resource_group_name,
-                managed_instance_name=managed_instance_name,
-                target_db_name=target_db_name,
-                parameters=parameters,
-                cls=lambda x,y,z: x,
-                **kwargs
-            )
-
-        kwargs.pop('error_map', None)
-        kwargs.pop('content_type', None)
-
-        def get_long_running_output(pipeline_response):
-            if cls:
-                return cls(pipeline_response, None, {})
-
-        path_format_arguments = {
-            'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'managedInstanceName': self._serialize.url("managed_instance_name", managed_instance_name, 'str'),
-            'targetDbName': self._serialize.url("target_db_name", target_db_name, 'str'),
-            'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
-        }
-
-        if polling is True: polling_method = AsyncARMPolling(lro_delay, path_format_arguments=path_format_arguments,  **kwargs)
-        elif polling is False: polling_method = AsyncNoPolling()
-        else: polling_method = polling
-        if cont_token:
-            return AsyncLROPoller.from_continuation_token(
-                polling_method=polling_method,
-                continuation_token=cont_token,
-                client=self._client,
-                deserialization_callback=get_long_running_output
-            )
-        else:
-            return AsyncLROPoller(self._client, raw_result, get_long_running_output, polling_method)
-    begin_cutover.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/managedInstances/{managedInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}/cutover'}  # type: ignore
+    begin_cancel.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/servers/{sqlDbInstanceName}/providers/Microsoft.DataMigration/databaseMigrations/{targetDbName}/cancel'}  # type: ignore