+++ src/datamigration/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.3.0
+++++++
+* [BREAKING CHANGES] Removing parameters provisioing-error and migration-operation-id from commands az datamigration sql-managed-instance create and az datamigration sql-vm create, as they are unnecessary parameters. 
+* [NEW SUBGROUP] az datamigration sql-db - Manage database migrations to SQL DB instance.
+
 0.2.0
 ++++++
 * Bug fix for Multiple connection strings in az datamigration get-assessment command.